% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltypes.R
\name{assign_cell_types}
\alias{assign_cell_types}
\title{Giorgio's functions for single cell sequencing analysis}
\usage{
assign_cell_types(so)
}
\arguments{
\item{so}{Seurat object}
}
\value{
Seurat object with additional columns 'SingleR.best_hit',
        'SingleR.best_score' and 'SingleR.best_match'
}
\description{
Acknowledgements:
some of the code was originally derived from code written by Sebastien Mella
}
\details{
Cell type assignment using SingleR
}
