% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggscrna.R
\name{assign_cell_types_V}
\alias{assign_cell_types_V}
\title{Verbose Cell type assignment using SingleR
(produces all outputs which were contained in the original code by S. Mella)}
\usage{
assign_cell_types_V(so)
}
\arguments{
\item{so}{Seurat object}
}
\value{
Seurat object with additional columns 'SingleR.best_hit',
        'SingleR.best_score' and 'SingleR.best_match'
}
\description{
Verbose Cell type assignment using SingleR
(produces all outputs which were contained in the original code by S. Mella)
}
