% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giosc.R
\name{boxplot_n_genes}
\alias{boxplot_n_genes}
\title{Boxplot of the number of genes per cell}
\usage{
boxplot_n_genes(so)
}
\arguments{
\item{so}{Seurat object}
}
\description{
Boxplot of the number of genes per cell
}
