% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggscrna.R
\name{boxplot_n_umis}
\alias{boxplot_n_umis}
\title{Boxplot of the number of UMIs per cell}
\usage{
boxplot_n_umis(so)
}
\arguments{
\item{so}{Seurat object}
}
\description{
Boxplot of the number of UMIs per cell
}
