% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellcycle.R
\name{cc_phase_plot}
\alias{cc_phase_plot}
\title{Create a plot of cell cycle phases}
\usage{
cc_phase_plot(so, sample = NULL, outfile = NULL)
}
\arguments{
\item{so}{Seurat object}

\item{sample}{Sample name (optional)
if provided: process only specified sample
default: consider all samples}

\item{outfile}{Output file (optional)
if provided: save the plot to the file
default: return the plot, but do not save it to file}
}
\value{
ggplot2 object
}
\description{
Create a plot of cell cycle phases
}
