% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellcycle.R
\name{cc_phase_table}
\alias{cc_phase_table}
\title{Compile a table of cell cycle phases}
\usage{
cc_phase_table(so, sample = NULL)
}
\arguments{
\item{so}{Seurat object}

\item{sample}{Sample name (optional) if the Seurat object contains
multiple samples, only the specified sample is considered;
default: consider all samples}
}
\value{
Character vector with the table
}
\description{
Compile a table of cell cycle phases
}
