% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giosc.R
\docType{data}
\name{CELLRANGER_FILTERED_COUNTS}
\alias{CELLRANGER_FILTERED_COUNTS}
\title{Functions for reading/merging/splitting a Seurat object.}
\format{
An object of class \code{character} of length 1.
}
\usage{
CELLRANGER_FILTERED_COUNTS
}
\description{
Acknowledgements:
some of the code was originally derived from code written by Sebastien Mella
}
\keyword{datasets}
