% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggscrna.R
\name{ComputeQCMetricsHuman}
\alias{ComputeQCMetricsHuman}
\title{Compute additional metrics on the Seurat object}
\usage{
ComputeQCMetricsHuman(so)
}
\arguments{
\item{so}{SeuratObject for which to compute the metrics}
}
\value{
The SeuratObject, with additional metrics in ``@meta.data``
}
\description{
The metrics are those explained in:
https://hbctraining.github.io/scRNA-seq/lessons/04_SC_quality_control.html
}
\details{
## Notes
- mt genes computation by MT- prefix assumes this is a human dataset,
  hence the name
}
