% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giosc.R
\name{considate_features}
\alias{considate_features}
\title{Consolidate multiple features of a Seurat object}
\usage{
considate_features(so, features_to_sum, new_feature_name)
}
\arguments{
\item{so}{Seurat object}

\item{features_to_sum}{Vector of strings. The names of the features to sum}

\item{new_feature_name}{String. The name of the new feature}
}
\value{
Seurat object with the specified features summed

Note: the function does not handle integrated data; re-run the integration
      after consolidating the features
}
\description{
Consolidate multiple features of a Seurat object
}
