% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggscrna.R
\name{Counts2Seurat}
\alias{Counts2Seurat}
\title{Read Counts using Read10X_h5 and call CreateSeuratObject}
\usage{
Counts2Seurat(
  sampleID,
  data_dir_template,
  counts_filename = "filtered_feature_bc_matrix.h5",
  min_cells = 0,
  min_features = 0
)
}
\arguments{
\item{sampleID}{String. Sample ID of the specific sample}

\item{data_dir_template}{String. Path to the sample directory, where the
sample ID is given as '{sampleID}'}

\item{counts_filename}{String. The filename containing the counts.}

\item{min_cells}{Integer. The min.cells parameter of CreateSeuratObject}

\item{min_features}{Integer. The min.features parameter of CreateSeuratObject}
}
\value{
SeuratObject from the given count matrix.
}
\description{
Streamlines reading from multiple directories under a common path
containing the output of "Cellranger count" for different samples
}
