% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giosc.R
\name{create_test_so}
\alias{create_test_so}
\title{Create a Seurat object for testing}
\usage{
create_test_so(
  num_genes,
  num_cells_per_sample,
  num_samples,
  low_expr_part = 0.8,
  low_avg = 2,
  high_avg = 20
)
}
\arguments{
\item{num_genes}{Integer. Number of genes.}

\item{num_cells_per_sample}{Integer. Number of cells per sample.}

\item{num_samples}{Integer. Number of samples.}

\item{low_expr_part}{Numeric. Proportion of genes with low expression.
(between 0 and 1)}

\item{low_avg}{Numeric. Average of low expression genes.
(used as lambda of Poisson distribution)}

\item{high_avg}{Numeric. Average of high expression genes.
(used as lambda of Poisson distribution)}
}
\value{
Seurat object.
}
\description{
Create a Seurat object for testing
}
