% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{density_plot_n_genes}
\alias{density_plot_n_genes}
\title{Plot number of genes per cell in different samples}
\usage{
density_plot_n_genes(so, plot_nrows = 0)
}
\arguments{
\item{so}{Seurat object}
}
\description{
Plot number of genes per cell in different samples
}
