% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggscrna.R
\name{density_plot_n_umis}
\alias{density_plot_n_umis}
\title{Plot number of UMIs per cell in different samples}
\usage{
density_plot_n_umis(so, plot_nrows = 0)
}
\arguments{
\item{so}{Seurat object}
}
\description{
Plot number of UMIs per cell in different samples
}
