% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{downsample_so}
\alias{downsample_so}
\title{Downsample a Seurat object to a given number of cells}
\usage{
downsample_so(so, ncells, verbose = TRUE, tic = TRUE)
}
\arguments{
\item{so}{Seurat object}

\item{ncells}{Number of cells to downsample to}

\item{verbose}{Print verbose messages}

\item{tic}{Print timing information}
}
\value{
Seurat object with the desired number of cells
        (or less if the original number of cells is less than ncells)
}
\description{
Downsample a Seurat object to a given number of cells
}
