% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giosc.R
\name{get_n_cells_order}
\alias{get_n_cells_order}
\title{Sample IDs ordered by number of cells}
\usage{
get_n_cells_order(so_list)
}
\arguments{
\item{so_list}{List of Seurat objects}
}
\value{
Vector of strings. Sample IDs in the order of the number of cells
}
\description{
Get the order of sample IDs according to the number of cells
in the Seurat objects
}
