% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_step_dir}
\alias{get_step_dir}
\title{Get the path to a step directory}
\usage{
get_step_dir(steps_dir, step_n)
}
\arguments{
\item{steps_dir}{String.  Path to the "steps" directory.}

\item{step_n}{String.  Prefix of the step.}
}
\value{
String.  Path to the step directory.
}
\description{
Get the path to a step directory
}
