% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggscrna.R
\name{GetSampleIDs}
\alias{GetSampleIDs}
\title{Get sample IDs from a sample sheet}
\usage{
GetSampleIDs(samples_sheet, column)
}
\arguments{
\item{samples_sheet}{String. Path to the TSV file}

\item{column}{String. Column containing the sample IDs}
}
\value{
Vector of strings. IDs of the samples.
}
\description{
The sample sheet must be in TSV format. The first line must be
the header, containing column names and optionally starting with
a '#' (without any spaces following it).
}
\examples{
samples <- GetSampleIDs("~/project/samples.tsv", "PatientID")

}
