% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggscrna.R
\name{h_expr_genes_plot}
\alias{h_expr_genes_plot}
\title{Plot the highest expressed genes}
\usage{
h_expr_genes_plot(so, n_genes = 30)
}
\arguments{
\item{so}{Seurat object}

\item{n_genes}{Integer. Number of genes to plot}
}
\value{
A ggplot object
}
\description{
based on a similar function by Sebastien Mella
}
