% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{histogram_n_cells}
\alias{histogram_n_cells}
\title{Plot number of cells in different samples}
\usage{
histogram_n_cells(so)
}
\arguments{
\item{so}{Seurat object}
}
\description{
Plot number of cells in different samples
}
