% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellcycle.R
\name{id_cc_phase}
\alias{id_cc_phase}
\title{Giorgio's functions for single cell sequencing analysis}
\usage{
id_cc_phase(so, genes = Seurat::cc.genes.updated.2019)
}
\arguments{
\item{so}{Seurat object}

\item{genes}{List of genes for cell cycle scoring}
}
\value{
Seurat object with additional columns 'Phase', 'S.Score'
        and 'G2M.Score'
}
\description{
Acknowledgements:
some of the code was originally derived from code written by Sebastien Mella
}
\details{
Identify cell cycle phases
}
