% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets.R
\name{id_geneset}
\alias{id_geneset}
\title{Identify genes in a gene set}
\usage{
id_geneset(so, geneset_name, verbose = TRUE, tic = FALSE)
}
\arguments{
\item{so}{a Seurat object}

\item{geneset_name}{a string, one of "TCR", "BCR", "MT", "RIBO"}

\item{verbose}{a boolean, if TRUE print the number of genes in the set}

\item{tic}{a boolean, if TRUE print the time taken to identify the genes}
}
\value{
a vector of gene names
}
\description{
Identify genes in a gene set
}
