% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggscrna.R
\name{MergeMultiSeurat}
\alias{MergeMultiSeurat}
\title{Merge multiple Seurat objects}
\usage{
MergeMultiSeurat(so_list, sampleIDs)
}
\arguments{
\item{so_list}{List of Seurat objects}

\item{sampleIDs}{List of strings, Sample IDs in the same order as in so_list}
}
\value{
A Seurat object.
    The cell IDs get the sample name prefixed to their barcode.
}
\description{
Merges all Seurat objects in the given list into a single object.
It is a wrapper for the merge function of SeuratObject.
}
