% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giosc.R
\name{multi_counts_to_seurat}
\alias{multi_counts_to_seurat}
\title{Create Seurat objects for multiple samples}
\usage{
multi_counts_to_seurat(
  sampleIDs,
  input_dir_template,
  counts_filename = CELLRANGER_FILTERED_COUNTS
)
}
\arguments{
\item{sampleIDs}{list of sampleIDs}

\item{input_dir_template}{String. Path to the sample directory, where the
sample ID is given as '{sampleID}'}

\item{counts_filename}{String. The filename containing the counts.}
}
\value{
list of Seurat objects for each of the samples
}
\description{
the sample IDs are added in the meta.data as 'sample' column
}
