% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggscrna.R
\name{qc_distri_plot}
\alias{qc_distri_plot}
\title{QC metric distribution histogram}
\usage{
qc_distri_plot(so, measure)
}
\arguments{
\item{so}{Seurat object}

\item{measure}{String. The name of the metric to plot}
}
\value{
A ggplot object
}
\description{
An histogram plot of the distribution
for one of the QC metrics in a Seurat object
}
