% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giosc.R
\name{read_samples_rds}
\alias{read_samples_rds}
\title{Read multiple Seurat Objects from RDS for each sample in a samples_sheet}
\usage{
read_samples_rds(
  samples_sheet,
  samples_sheet_idcol,
  file_pfx,
  file_sfx,
  verbose = FALSE
)
}
\arguments{
\item{samples_sheet}{String. Path to the samples sheet}

\item{samples_sheet_idcol}{String. Column name of the sample ID in the sheet}

\item{file_pfx}{String. Prefix of the RDS files}

\item{file_sfx}{String. Suffix of the RDS files}

\item{verbose}{(optional) Boolean. Whether to print the sample IDs
Defaults to FALSE}
}
\value{
List of Seurat objects
}
\description{
Read multiple Seurat Objects from RDS for each sample in a samples_sheet
}
