% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples_sheet.R
\name{read_samples_sheet}
\alias{read_samples_sheet}
\title{Handling of the samples sheet data}
\usage{
read_samples_sheet(filename, column = "SampleID", verbose = TRUE)
}
\arguments{
\item{column}{String. Column containing the sample IDs}

\item{verbose}{Boolean. Whether to print the sample sheet and the IDs.}

\item{samples_sheet}{String. Path to the TSV file}
}
\value{
data.frame. The sample sheet.
}
\description{
Read samples sheet
}
\details{
The sample sheet must be in TSV format. The first line must be
the header, containing column names and optionally starting with
a '#' (without any spaces following it).
}
\examples{
  samples_sheet <- read_samples_sheet("~/project/samples.tsv", "PatientID")

}
