% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets.R
\name{rm_geneset}
\alias{rm_geneset}
\title{Remove gene set from a Seurat object}
\usage{
rm_geneset(so, geneset_name, verbose = TRUE, tic = FALSE)
}
\arguments{
\item{so}{the Seurat object}

\item{geneset_name}{String. Geneset name
it must be one of the elements of 'genesets'}

\item{verbose}{Boolean. TRUE: be verbose}

\item{tic}{Boolean. TRUE: show the running time}
}
\value{
the Seurat object without the genes in the specified geneset
}
\description{
Remove gene set from a Seurat object
}
