% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets.R
\name{rm_genesets}
\alias{rm_genesets}
\title{Remove multiple genesets based on a list of parameters}
\usage{
rm_genesets(so, params, verbose = TRUE, tic = TRUE)
}
\arguments{
\item{so}{a Seurat object}

\item{params}{parameters, either a list or environment.
"RIBO" and the value is 'rm', then the corresponding gene set is
removed}
}
\description{
Remove multiple genesets based on a list of parameters
}
