% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giosc.R
\name{robustbase_qc}
\alias{robustbase_qc}
\title{Add QC information from robustbase library to Seurat object}
\usage{
robustbase_qc(so)
}
\arguments{
\item{so}{Seurat object}
}
\value{
Seurat object with additional QC metrics
}
\description{
Add QC information from robustbase library to Seurat object
}
