% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{scater_qc}
\alias{scater_qc}
\title{Add QC information from scater library to Seurat object}
\usage{
scater_qc(so)
}
\arguments{
\item{so}{Seurat object}
}
\value{
Seurat object with additional QC metrics
}
\description{
Add QC information from scater library to Seurat object
}
