% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggscrna.R
\name{SeuratReadMulti}
\alias{SeuratReadMulti}
\title{Create Seurat objects for multiple samples}
\usage{
SeuratReadMulti(
  sampleIDs,
  data_dir_template,
  counts_filename = "filtered_feature_bc_matrix.h5"
)
}
\arguments{
\item{sampleIDs}{list of sampleIDs}

\item{data_dir_template}{String. Path to the sample directory, where the
sample ID is given as '{sampleID}'}

\item{counts_filename}{String. The filename containing the counts.}
}
\value{
list of Seurat objects for each of the samples
}
\description{
Creates multiple Seurat objects for a list of samples
and adds further QC metrics to them
}
