% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggscrna.R
\name{show_qc_plots}
\alias{show_qc_plots}
\title{Create and print several QC metrics plots}
\usage{
show_qc_plots(so, plot_nrows = 0)
}
\arguments{
\item{so}{Seurat object}
}
\description{
Creates several QC metrics plots for a Seurat object
and prints them to the screen
}
