% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giosc.R
\name{so_extract_sample}
\alias{so_extract_sample}
\title{Check that a sample exists and extract it}
\usage{
so_extract_sample(so, sample)
}
\arguments{
\item{so}{Seurat object}

\item{samplename}{String. Sample name}
}
\value{
Seurat object with only the cells of the specified sample
}
\description{
Check that a sample exists and extract it
}
