% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giosc.R
\name{assign_cell_types}
\alias{assign_cell_types}
\title{Cell type assignment using SingleR}
\usage{
assign_cell_types(so)
}
\arguments{
\item{so}{Seurat object}
}
\value{
Seurat object with additional columns 'SingleR.best_hit',
        'SingleR.best_score' and 'SingleR.best_match'
}
\description{
Cell type assignment using SingleR
}
