% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellcycle.R
\name{cc_phase_plot}
\alias{cc_phase_plot}
\title{Create a plot of cell cycle phases}
\usage{
cc_phase_plot(so, sample = NULL)
}
\arguments{
\item{so}{Seurat object}

\item{sample}{Sample name (optional) if the Seurat object contains
multiple samples, only the specified sample is considered;
default: consider all samples}
}
\value{
ggplot2 object
}
\description{
Create a plot of cell cycle phases
}
