% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{h_expr_genes_plot}
\alias{h_expr_genes_plot}
\title{Plot the highest expressed genes}
\usage{
h_expr_genes_plot(so, n_genes = 10, sample = NULL)
}
\arguments{
\item{so}{Seurat object}

\item{n_genes}{Integer. Number of genes to plot}

\item{sample}{String. Keep only specified sample. Default: keep all.}
}
\value{
A ggplot object
}
\description{
Plot the highest expressed genes
}
