% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giosc.R
\name{id_cc_phase}
\alias{id_cc_phase}
\title{Identify cell cycle phases}
\usage{
id_cc_phase(so, genes = cc.genes.updated.2019)
}
\arguments{
\item{so}{Seurat object}

\item{genes}{List of genes for cell cycle scoring}
}
\value{
Seurat object with additional columns 'Phase', 'S.Score'
        and 'G2M.Score'
}
\description{
derived from a function of Sebastien Mella
}
