% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets.R
\name{id_geneset}
\alias{id_geneset}
\title{Identify genes in a gene set}
\usage{
id_geneset(so, geneset_name, verbose = TRUE, tic = FALSE)
}
\arguments{
\item{so}{the Seurat object}

\item{geneset_name}{String. Geneset name
it must be one of the elements of 'genesets'}

\item{verbose}{Boolean. TRUE: be verbose}

\item{tic}{Boolean. TRUE: show the running time}
}
\value{
Vector. Gene names
}
\description{
Identify genes in a gene set
}
