% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets.R
\name{rm_geneset}
\alias{rm_geneset}
\title{Remove gene set from a Seurat object}
\usage{
rm_geneset(so, geneset_name, verbose = TRUE, tic = FALSE)
}
\arguments{
\item{so}{a Seurat object}

\item{geneset_name}{a string, one of "TCR", "BCR", "MT", "RIBO"}

\item{verbose}{a boolean, if TRUE print the number of genes in the set}

\item{tic}{a boolean, if TRUE print the time taken to remove the genes}
}
\value{
a Seurat object
}
\description{
Remove gene set from a Seurat object
}
