% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets.R
\name{rm_genesets}
\alias{rm_genesets}
\title{Remove multiple genesets based on a list of parameters}
\usage{
rm_genesets(so, params, verbose = TRUE, tic = TRUE)
}
\arguments{
\item{so}{the Seurat object}

\item{params}{List or Environment. Parameters}

\item{verbose}{Boolean. TRUE: be verbose}

\item{tic}{Boolean. TRUE: show the running time}
}
\value{
the Seurat object without the genes in the specified genesets
}
\description{
It calls rm_geneset for each geneset in the parameters list or env,
where a parameter is called after a geneset and its value is 'rm'
}
