% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{source_rmd}
\alias{source_rmd}
\title{Helper functions for the giosc package}
\usage{
source_rmd(file, ...)
}
\arguments{
\item{file}{Filename of the file to be included}
}
\description{
Include a Rmd file in another
}
